requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
asyncer==0.0.8
attrs==24.3.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
coloredlogs==15.0.1
colorlog==6.9.0
dataclasses-json==0.6.7
datasets==3.2.0
decorator==5.1.1
deprecated==1.2.15
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
dspy==2.5.42
duckduckgo-search==7.1.1
durationpy==0.9
email-validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.9.0.post1
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.37.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json-repair==0.35.0
jsonpickle==4.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
litellm==1.51.0
llama-cloud==0.1.7
llama-index==0.12.8
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.8
llama-index-embeddings-ollama==0.5.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-ollama==0.5.0
llama-index-llms-openai==0.3.12
llama-index-multi-modal-llms-openai==0.4.1
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-database==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-github==0.5.0
llama-index-readers-llama-parse==0.4.0
llama-index-utils-workflow==0.3.0
llama-index-vector-stores-chroma==0.4.1
llama-index-vector-stores-faiss==0.3.0
llama-parse==0.5.19
loguru==0.7.3
lxml==5.3.0
magicattr==0.1.6
mako==1.3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.23.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
oauthlib==3.2.2
ollama==0.4.5
onnxruntime==1.20.1
openai==1.58.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
optuna==4.1.0
orjson==3.10.13
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
posthog==3.7.4
primp==0.9.2
prompt-toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pygments==2.18.0
pypdf==5.1.0
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyventus==0.6.0
pyvis==0.3.2
pyyaml==6.0.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
rsa==4.9
ruff==0.8.4
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.15.0
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.36
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.41.3
striprtf==0.0.26
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.2.1
torch==2.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
tree-sitter==0.23.2
tree-sitter-languages==1.10.2
typer==0.15.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
ujson==5.10.0
urllib3==2.3.0
uuid==1.30
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
